refactor: moved the corresponding files to the ask/__components
