+++ src/ui/Ask/Ask.tsx
@@ -1,7 +1,7 @@
 import { useTrans } from 'ui/_hooks/useTrans'
-import { Header, Form, SectionTitle } from 'ui/_components'
+import { Header, SectionTitle } from 'ui/_components'
+import { Form } from './__components/Form'
 import styled from '@emotion/styled'
-import { colors } from 'ui/_styles'
 import React from 'react'
 import * as yup from 'yup'
 import { Formik } from 'formik'
