+++ src/ui/_components/index.ts
@@ -1,3 +1,2 @@
 export { DefaultFormElementWrapper, Box, Row } from './objects/'
 export { SectionTitle, Header } from './molecules'
-export { Form } from './organisms'
