+++ src/ui/_components/molecules/index.ts
@@ -1,7 +1,4 @@
-export { SmallFormInput } from './inputs/SmallFormInput'
-export { BigFormInput } from './inputs/BigFormInput'
 export { SelectBox } from './inputs/SelectBox'
-export { FormButton } from './buttons/FormButton'
 export { Header } from './titles/Header'
 export { SectionTitle } from './titles/SectionTitle'
 export { TalkCard } from './cards/TalkCard'
