--- src/ui/_components/organisms/index.ts
@@ -1,1 +1,0 @@
-export { Form } from './Form'
